+++ src/main/java/ar/edu/unju/fi/tp9/repository/MiembroRepository.java
@@ -11,13 +11,13 @@ import ar.edu.unju.fi.tp9.entity.Miembro;
 
 
 @Repository
-public interface MiembroRepository extends CrudRepository<Miembro, Integer>{
+public interface MiembroRepository extends CrudRepository<Miembro, Long>{
     
     public Optional<Miembro> findByNombre(String nombre);
 
     public Optional<Miembro> findByCorreo(String correo);
 
-    public void deleteById(Integer id);
+    public void deleteById(Long id);
 
     
 }
